use int instead of long

+++ src/ipv6cp.h
@@ -32,7 +32,7 @@
     u_char 		myintid[8];
     u_char 		hisintid[8];
 
-    u_long		peer_reject;	/* Request codes rejected by peer */
+    u_int		peer_reject;	/* Request codes rejected by peer */
 
     struct fsm		fsm;
   };
